03- Mastering OpenAI APIs/05- Youtube Summary/requirements.txt
openai
python-dotenv
fastapi==0.95.1
uvicorn==0.20.0
pytube
python-multipart